requirements.txt
torch==2.5.1
torchvision==0.16.1
torchaudio==2.5.1
transformers==4.35.0
pillow==10.0.0
